QuarkBuilder 0.4.0 (QuotePay support)  (#13)
Introduces `QuotePay` to the QuarkBuilder. Rather than wrapping operations around a Paycall/Quotecall to pay for operations, a `QuotePay` is used instead. `QuotePay` takes an amount (e.g. quote) and transfers that from the user to a payee. A `QuotePay` can be made on any chain that the user has enough payment assets on, which means we no longer have to deal with the complexities of bridging payment tokens.

The quote that the `QuotePay` uses is passed in by the client. Currently, only the `transfer` builder function supports this new Quote interface. We plan to update the interfaces for the other builder functions as well in a follow up PR.

---------

Co-authored-by: Mykel Pereira <me@mykelpereira.com>