Acceptance Tests
This is a feature branch for Acceptance Tests, which are tests that test externally observable behaviors of QuarkBuilder. These allow us to (a) document, at a high level, the behaviors of Quark Builder, (b) act as regression tests, and (c) allow larger refactors of the code.

Please merge down all Acceptance Tests PRs into this branch until this is merged to main.