Fix Bridge Token Accounting for Payment
This patch fixes Bridge Token accouting for payment. Basically, we were incorrectly accouting for asset movement of the Payment token over Across. Specifically, we would look at the bridge action and say "We are moving WETH from L1 to L2" and then when we looked at the wrap action we would say "and we get more from wrapping it, too!". This patch fixes it so that for WETH payment token we properly discount the L1 balance and only single count the L2 balance.

We should really consider making the trackAssetsInAndOut function more generic, since it's specific to payment token and might have more of these quirks.

Patches:
  * Simplify for the logic we're specifically looking for